monitoring/requirements.txt
pika
psycopg2-binary
fastapi
uvicorn
websockets
